finish the simple framework
getMeasureWidth() getWidth()
getWidth() is after layout ,otherwite it returns 0;

getX() is getLeft() + getTranslationX();
ogetLeft() returns the views location relative to its parent. How it has scrolled does not impact this at all. Scrolling impacts the contents of the view, not its location.
a quote from the android documentation regarding getLeft():

For instance, when getLeft() returns 20, that means the view is located 20 pixels to
the right of the left edge of its direct parent.
getScrollX(), on the other hand, lets you know how the content in the view has moved.

View.scrollBy(20,0) affects the content in the view (like subviews of the view) and doesn't actually move the view relative to the view's parent.
getTranslaitonX () :The horizontal location of this view relative to its left position. This position is post-layout, in addition to wherever the object's layout placed it.